@@ -1,9 +1,11 @@
 module github.com/launchdarkly/go-test-helpers/v2
 
-go 1.13
+go 1.18
+
+require github.com/stretchr/testify v1.5.1
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/stretchr/testify v1.5.1
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v2 v2.2.8 // indirect
 )
